ml-fastapi/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1 
requests==2.18.4
typing==3.7.4.3
tensorflow-macos==2.8.0
tensorflow-text==2.8.2
pydantic==1.10.6
PyMuPDF==1.22.1
googletrans==3.1.0a0